+++ packages/communication/src/wallet/channels.ts
@@ -8,7 +8,6 @@ export type UserInfo = {
   typeOfLogin: string;
   verifier: string;
   verifierId: string;
-  idToken?: string;
   isNewUser?: boolean;
 };