+++ packages/embed-wallet/src/constants.ts
@@ -1,1 +1,1 @@
-export const WALLET_CLIENT_VERSION = '1.25.0-rc.1';
+export const WALLET_CLIENT_VERSION = '1.25.0';