+++ packages/wallet-engine/src/engine/provider.ts
@@ -31,20 +31,19 @@ class EngineProvider extends EventEmitter implements Provider {
   }
 }
 
-export class ProviderEngine extends Engine {
+class UnsafeEngine extends Engine {
   readonly provider: Provider = new EngineProvider(this);
 
-  constructor(private options: ProviderEngineOptions) {
+  constructor(options: ProviderEngineOptions) {
     super();
 
     this.pushEngine(createWalletEngine(options));
-    this.pushEngine(createApproveEngine(options));
 
-    this.pushEngine(async () => {
-      const { createAccountsEngine } = await import(/* webpackChunkName: "accountsEngine" */ './accounts');
-
-      return createAccountsEngine(options);
-    });
+    this.pushEngine(
+      import(/* webpackChunkName: "accountsEngine" */ './accounts').then(({ createAccountsEngine }) =>
+        createAccountsEngine(options),
+      ),
+    );
 
     this.pushEngine(async () => {
       const { createPolkadotEngine } = await import(/* webpackChunkName: "polkadotEngine" */ './polkadot');
@@ -61,6 +60,23 @@ export class ProviderEngine extends Engine {
 
       return createEthereumEngine(options);
     });
+  }
+}
+
+export class ProviderEngine extends Engine {
+  readonly provider: Provider;
+  readonly unsafeProvider;
+
+  constructor(options: ProviderEngineOptions) {
+    super();
+
+    const unsafeEngine = new UnsafeEngine(options);
+
+    this.provider = new EngineProvider(this);
+    this.unsafeProvider = new EngineProvider(unsafeEngine);
+
+    this.pushEngine(createApproveEngine(options));
+    this.pushEngine(unsafeEngine);
   }
 
   async updateAccounts() {