+++ src/routes/Authorize/AuthorizeRedirect.tsx
@@ -1,44 +1,26 @@
 import { observer } from 'mobx-react-lite';
 import { Loading, Logo, Alert } from '@cere-wallet/ui';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
-import { OpenLoginStore } from '~/stores/OpenLoginStore';
-
-const createRedirectUrl = (url: string, sessionId: string) => {
-  const finalUrl = new URL(url);
-  const hashParams = new URLSearchParams(finalUrl.hash.slice(1));
-
-  hashParams.append('sessionId', sessionId);
-  finalUrl.hash = hashParams.toString();
-
-  return finalUrl.toString();
-};
+import { useOutletContext } from 'react-router-dom';
+import { AuthorizePopupStore } from '~/stores';
 
 const AuthorizeRedirect = () => {
   const [error, setError] = useState<string>();
   const hashParams = new URLSearchParams(window.location.hash.slice(1));
   const queryParams = new URLSearchParams(window.location.search);
-  const sessionId = hashParams.get('sessionId');
   const result = hashParams.get('result');
-  const sessionNamespace = hashParams.get('sessionNamespace') || undefined;
   const redirectUrl = queryParams.get('redirectUrl');
 
-  const store = useMemo(() => new OpenLoginStore({ sessionNamespace }), [sessionNamespace]);
+  const store = useOutletContext<AuthorizePopupStore>();
 
   const handleAuthResult = useCallback(async () => {
-    if (!result || !sessionId || !redirectUrl) {
+    if (!result || !redirectUrl) {
       return;
     }
 
-    const isRedirectAllowed = await store.isAllowedRedirectUrl(redirectUrl);
-
-    if (!isRedirectAllowed) {
-      throw new Error('The redirect url is not allowed');
-    }
-
-    store.syncWithEncodedState(result, sessionId);
-    window.location.replace(createRedirectUrl(redirectUrl, sessionId));
-  }, [store, redirectUrl, result, sessionId]);
+    await store.acceptEncodedState(result);
+  }, [store, redirectUrl, result]);
 
   useEffect(() => {
     handleAuthResult().catch((error: Error) => setError(error.message));