+++ src/stores/Web3AuthStore/index.ts
@@ -1,0 +1,2 @@
+export * from './Web3AuthStore';
+export * from './getScopedKey';