+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ### vNext
 
 - [[Feature](https://www.notion.so/cere/Switch-DEV-frontends-to-new-infra-563be1f2c5ee4b92aa87808f3480fe22?pvs=4)] Switch URL to new infra
+
+### v1.25.1
+
+- Fix authentication retry when it shows previous error without actual retrying
+
+### v1.25.0
+
+- [[Feature](https://www.notion.so/cere/Cere-Wallet-simulation-config-46f78c7d573e44f198eac83c1805e482?pvs=4)] Add simulation workflow to run critical user path
+- [[Feature](https://www.notion.so/cere/Plan-for-TORUS-replacement-1304aeda9f8c468b946b8b223fb86da7?pvs=4)] Update authentication flow to reduce new users onboarding time
+- [[Feature](https://www.notion.so/cere/Implement-E2E-UI-tests-for-Wallet-Client-to-check-Tor-us-fix-b6575b5474684b6cb1566926063a9254?pvs=4)] Add E2E UI tests configuration and simple login spec
 - [[Fix](https://www.notion.so/cere/Missed-version-in-Cere-Wallet-IFRAME-URL-on-Stage-c647d3be181846aeb8f6d7fd908ab385?pvs=4)] Add version to the wallet IFRAME URL on Stage environment
 - [[Feature](https://www.notion.so/cere/Whitehat-Wallet-Open-redirect-bypass-89bba775a9fc4db58ecbaec5a9935268)] Validate that login redirect URL is whitelisted to prevent fishing attacks